+++ src/Rsse.Service/Api/Controllers/CreateController.cs
@@ -71,7 +71,6 @@ public class CreateController(
 
             var path = CreateDumpAndGetFilePath();
 
-            // todo: можно добавить в маппер
             noteResultDto = noteResultDto with
             {
                 Text = path ?? string.Empty